+++ amplify/src/main/java/com/github/stkent/amplify/tracking/Amplify.java
@@ -27,7 +27,6 @@ import com.github.stkent.amplify.prompt.PromptPresenter;
 import com.github.stkent.amplify.prompt.interfaces.IPromptPresenter;
 import com.github.stkent.amplify.prompt.interfaces.IPromptView;
 import com.github.stkent.amplify.tracking.interfaces.IAppEventTimeProvider;
-import com.github.stkent.amplify.tracking.interfaces.IAppInfoProvider;
 import com.github.stkent.amplify.tracking.interfaces.IAppLevelEventRulesManager;
 import com.github.stkent.amplify.tracking.interfaces.IEnvironmentBasedRule;
 import com.github.stkent.amplify.tracking.interfaces.IEnvironmentBasedRulesManager;
@@ -39,15 +38,17 @@ import com.github.stkent.amplify.tracking.managers.AppLevelEventRulesManager;
 import com.github.stkent.amplify.tracking.managers.EnvironmentBasedRulesManager;
 import com.github.stkent.amplify.tracking.managers.FirstEventTimeRulesManager;
 import com.github.stkent.amplify.tracking.managers.LastEventTimeRulesManager;
-import com.github.stkent.amplify.tracking.managers.LastEventVersionRulesManager;
+import com.github.stkent.amplify.tracking.managers.LastEventVersionCodeRulesManager;
+import com.github.stkent.amplify.tracking.managers.LastEventVersionNameRulesManager;
 import com.github.stkent.amplify.tracking.managers.TotalEventCountRulesManager;
-import com.github.stkent.amplify.tracking.prerequisites.GooglePlayStoreRule;
+import com.github.stkent.amplify.tracking.rules.GooglePlayStoreRule;
 import com.github.stkent.amplify.tracking.rules.MaximumCountRule;
-import com.github.stkent.amplify.tracking.rules.VersionChangedRule;
+import com.github.stkent.amplify.tracking.rules.VersionNameChangedRule;
 import com.github.stkent.amplify.utils.ActivityStateUtil;
-import com.github.stkent.amplify.utils.AppInfoProvider;
 import com.github.stkent.amplify.utils.FeedbackUtil;
 import com.github.stkent.amplify.utils.PlayStoreUtil;
+import com.github.stkent.amplify.utils.appinfo.AppInfoUtil;
+import com.github.stkent.amplify.utils.appinfo.IAppInfoProvider;
 
 public final class Amplify implements IEventListener {
 
@@ -61,7 +62,8 @@ public final class Amplify implements IEventListener {
     private final IEnvironmentBasedRulesManager environmentBasedRulesManager;
     private final FirstEventTimeRulesManager firstEventTimeRulesManager;
     private final LastEventTimeRulesManager lastEventTimeRulesManager;
-    private final LastEventVersionRulesManager lastEventVersionRulesManager;
+    private final LastEventVersionCodeRulesManager lastEventVersionCodeRulesManager;
+    private final LastEventVersionNameRulesManager lastEventVersionNameRulesManager;
     private final TotalEventCountRulesManager totalEventCountRulesManager;
 
     private final ILogger logger;
@@ -87,8 +89,8 @@ public final class Amplify implements IEventListener {
     // constructors
 
     private Amplify(@NonNull final Context context, @NonNull final ILogger logger) {
-        AppInfoProvider.initialize(context);
-        final IAppInfoProvider appInfoProvider = AppInfoProvider.getSharedInstance();
+        AppInfoUtil.initialize(context);
+        final IAppInfoProvider appInfoProvider = AppInfoUtil.getSharedAppInfoProvider();
 
         final Context appContext = context.getApplicationContext();
         final IAppEventTimeProvider appEventTimeProvider
@@ -104,9 +106,14 @@ public final class Amplify implements IEventListener {
                 = new EnvironmentBasedRulesManager(environmentCapabilitiesProvider, logger);
 
         this.firstEventTimeRulesManager = new FirstEventTimeRulesManager(appContext, logger);
+
         this.lastEventTimeRulesManager = new LastEventTimeRulesManager(appContext, logger);
-        this.lastEventVersionRulesManager
-                = new LastEventVersionRulesManager(appContext, appInfoProvider, logger);
+
+        this.lastEventVersionNameRulesManager
+                = new LastEventVersionNameRulesManager(appContext, appInfoProvider, logger);
+
+        this.lastEventVersionCodeRulesManager
+                = new LastEventVersionCodeRulesManager(appContext, appInfoProvider, logger);
 
         this.totalEventCountRulesManager = new TotalEventCountRulesManager(appContext, logger);
 
@@ -127,12 +134,12 @@ public final class Amplify implements IEventListener {
                 .setLastCrashTimeCooldownDays(DEFAULT_LAST_CRASH_TIME_COOLDOWN_DAYS)
                 .addTotalEventCountRule(PromptViewEvent.USER_GAVE_POSITIVE_FEEDBACK,
                         new MaximumCountRule(DEFAULT_USER_GAVE_POSITIVE_FEEDBACK_MAXIMUM_COUNT))
-                .addLastEventVersionRule(PromptViewEvent.USER_GAVE_CRITICAL_FEEDBACK,
-                        new VersionChangedRule())
-                .addLastEventVersionRule(PromptViewEvent.USER_DECLINED_CRITICAL_FEEDBACK,
-                        new VersionChangedRule())
-                .addLastEventVersionRule(PromptViewEvent.USER_DECLINED_POSITIVE_FEEDBACK,
-                        new VersionChangedRule());
+                .addLastEventVersionNameRule(PromptViewEvent.USER_GAVE_CRITICAL_FEEDBACK,
+                        new VersionNameChangedRule())
+                .addLastEventVersionNameRule(PromptViewEvent.USER_DECLINED_CRITICAL_FEEDBACK,
+                        new VersionNameChangedRule())
+                .addLastEventVersionNameRule(PromptViewEvent.USER_DECLINED_POSITIVE_FEEDBACK,
+                        new VersionNameChangedRule());
     }
 
     public Amplify addEnvironmentBasedRule(@NonNull final IEnvironmentBasedRule rule) {
@@ -179,11 +186,19 @@ public final class Amplify implements IEventListener {
         return this;
     }
 
-    public Amplify addLastEventVersionRule(
+    public Amplify addLastEventVersionCodeRule(
+            @NonNull final IEvent event,
+            @NonNull final IEventBasedRule<Integer> rule) {
+
+        lastEventVersionCodeRulesManager.addEventBasedRule(event, rule);
+        return this;
+    }
+
+    public Amplify addLastEventVersionNameRule(
             @NonNull final IEvent event,
             @NonNull final IEventBasedRule<String> rule) {
 
-        lastEventVersionRulesManager.addEventBasedRule(event, rule);
+        lastEventVersionNameRulesManager.addEventBasedRule(event, rule);
         return this;
     }
 
@@ -212,7 +227,8 @@ public final class Amplify implements IEventListener {
         totalEventCountRulesManager.notifyEventTriggered(event);
         firstEventTimeRulesManager.notifyEventTriggered(event);
         lastEventTimeRulesManager.notifyEventTriggered(event);
-        lastEventVersionRulesManager.notifyEventTriggered(event);
+        lastEventVersionCodeRulesManager.notifyEventTriggered(event);
+        lastEventVersionNameRulesManager.notifyEventTriggered(event);
     }
 
     // Query methods
@@ -252,7 +268,7 @@ public final class Amplify implements IEventListener {
                         PlayStoreUtil.openPlayStoreToRate(activity, packageName);
                     } else if (event == PromptViewEvent.USER_GAVE_CRITICAL_FEEDBACK) {
                         final IAppInfoProvider appInfoProvider
-                                = AppInfoProvider.getSharedInstance();
+                                = AppInfoUtil.getSharedAppInfoProvider();
 
                         final FeedbackUtil feedbackUtil = new FeedbackUtil(
                                 new AppFeedbackDataProvider(appInfoProvider),
@@ -281,7 +297,8 @@ public final class Amplify implements IEventListener {
                 & totalEventCountRulesManager.shouldAllowFeedbackPrompt()
                 & firstEventTimeRulesManager.shouldAllowFeedbackPrompt()
                 & lastEventTimeRulesManager.shouldAllowFeedbackPrompt()
-                & lastEventVersionRulesManager.shouldAllowFeedbackPrompt());
+                & lastEventVersionCodeRulesManager.shouldAllowFeedbackPrompt()
+                & lastEventVersionNameRulesManager.shouldAllowFeedbackPrompt());
     }
 
 }