+++ amplify/src/main/java/com/github/stkent/amplify/tracking/rules/GooglePlayStoreRule.java
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.github.stkent.amplify.tracking.prerequisites;
+package com.github.stkent.amplify.tracking.rules;
 
 import android.support.annotation.NonNull;
 
@@ -25,7 +25,7 @@ import com.github.stkent.amplify.tracking.interfaces.IEnvironmentCapabilitiesPro
  * An implementation of {@code IEnvironmentBasedRule} that verifies whether or not
  * the Google Play Store is installed on the current device.
  */
-public class GooglePlayStoreRule implements IEnvironmentBasedRule {
+public final class GooglePlayStoreRule implements IEnvironmentBasedRule {
 
     @Override
     public boolean shouldAllowFeedbackPrompt(@NonNull final IEnvironmentCapabilitiesProvider environmentInfoProvider) {