+++ update_toc.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+doctoc README.md --github --maxlevel 2